# frontend/docs-backend/openapi_QUYNH_v8.0.0.yaml
...
openapi: 3.1.0
info:
  title: Quynh Web API
  description: Cổng thông tin điện tử nội bộ Lữ đoàn Thông tin 21. Xem lịch sử thay
    đổi ở openapi.CHANGELOG.md.
  version: 8.0.0
paths:
...
  /directives:
    post:
      tags:
      - directives
      summary: Create Directive
      operationId: create_directive_directives_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectiveCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectiveOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - directives
      summary: List Directives
      operationId: list_directives_directives_get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Filter
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectiveOut'
                title: Response List Directives Directives Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /directives/{directive_id}:
    get:
      tags:
      - directives
      summary: Get Directive
      operationId: get_directive_directives__directive_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: directive_id
        in: path
        required: true
        schema:
          type: integer
          title: Directive Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectiveOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - directives
      summary: Update Directive
      operationId: update_directive_directives__directive_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: directive_id
        in: path
        required: true
        schema:
          type: integer
          title: Directive Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectiveCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectiveOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - directives
      summary: Delete Directive
      operationId: delete_directive_directives__directive_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: directive_id
        in: path
        required: true
        schema:
          type: integer
          title: Directive Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /directives/{directive_id}/acknowledge:
    post:
      tags:
      - directives
      summary: Acknowledge Directive
      operationId: acknowledge_directive_directives__directive_id__acknowledge_post
      security:
      - HTTPBearer: []
      parameters:
      - name: directive_id
        in: path
        required: true
        schema:
          type: integer
          title: Directive Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectiveOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /directives/{directive_id}/acknowledgements:
    get:
      tags:
      - directives
      summary: Get Acknowledgement Report
      operationId: get_acknowledgement_report_directives__directive_id__acknowledgements_get
      security:
      - HTTPBearer: []
      parameters:
      - name: directive_id
        in: path
        required: true
        schema:
          type: integer
          title: Directive Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectiveAckReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /chats/direct:
    post:
      tags:
      - chats
      summary: Open Or Create Direct Chat
      description: Mở hoặc khởi tạo cuộc trò chuyện trực tiếp 1-1 với một đồng chí.
      operationId: open_or_create_direct_chat_chats_direct_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectChatCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatConversationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /contact-books:
    get:
      tags:
      - contacts
      summary: List Books
      operationId: list_books_contact_books_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContactBookOut'
                type: array
                title: Response List Books Contact Books Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - contacts
      summary: Import Book
      operationId: import_book_contact_books_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_book_contact_books_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactBookOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /contact-books/{book_id}:
    get:
      tags:
      - contacts
      summary: Get Book
      operationId: get_book_contact_books__book_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactBookOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - contacts
      summary: Delete Book
      operationId: delete_book_contact_books__book_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /contact-books/{book_id}/contacts:
    get:
      tags:
      - contacts
      summary: List Contacts
      operationId: list_contacts_contact_books__book_id__contacts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Từ khoá tra cứu trên mọi trường
          title: Q
        description: Từ khoá tra cứu trên mọi trường
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_import_book_contact_books_post:
      properties:
        name:
          type: string
          title: Name
          default: ''
        description:
          type: string
          title: Description
          default: ''
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_import_book_contact_books_post
...
    ChatConversationOut:
      properties:
        id:
          type: integer
          title: Id
        type:
          type: string
          title: Type
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        created_by_id:
          type: integer
          title: Created By Id
        created_at:
          type: string
          format: date-time
          title: Created At
        last_message_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Message At
        last_message_preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Message Preview
        unread_count:
          type: integer
          title: Unread Count
          default: 0
        status:
          type: string
          title: Status
          default: da_duyet
        review_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Note
        reviewed_by_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reviewed By Id
        participants:
          items:
            $ref: '#/components/schemas/ChatParticipantOut'
          type: array
          title: Participants
          default: []
        other_user:
          anyOf:
          - $ref: '#/components/schemas/ChatParticipantOut'
          - type: 'null'
      type: object
      required:
      - id
      - type
      - created_by_id
      - created_at
      title: ChatConversationOut
...
    ChatParticipantOut:
      properties:
        user_id:
          type: integer
          title: User Id
        username:
          type: string
          title: Username
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        rank:
          anyOf:
          - type: string
          - type: 'null'
          title: Rank
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        unit_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unit Id
        unit_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Name
        is_admin:
          type: boolean
          title: Is Admin
          default: false
        joined_at:
          type: string
          format: date-time
          title: Joined At
      type: object
      required:
      - user_id
      - username
      - joined_at
      title: ChatParticipantOut
...
    ContactBookOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        source_file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File Name
        column_headers:
          items:
            type: string
          type: array
          title: Column Headers
        row_count:
          type: integer
          title: Row Count
        created_by_id:
          type: integer
          title: Created By Id
        created_by_full_name:
          type: string
          title: Created By Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - description
      - source_file_name
      - column_headers
      - row_count
      - created_by_id
      - created_by_full_name
      - created_at
      title: ContactBookOut
    ContactOut:
      properties:
        id:
          type: integer
          title: Id
        book_id:
          type: integer
          title: Book Id
        row_index:
          type: integer
          title: Row Index
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        extra:
          additionalProperties:
            type: string
          type: object
          title: Extra
      type: object
      required:
      - id
      - book_id
      - row_index
      - full_name
      - unit
      - position
      - phone
      - email
      - extra
      title: ContactOut
    ContactPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ContactOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
        columns:
          items:
            type: string
          type: array
          title: Columns
      type: object
      required:
      - items
      - total
      - skip
      - limit
      - columns
      title: ContactPage
      description: Ket qua phan trang cho GET /contact-books/{id}/contacts.
    DirectChatCreate:
      properties:
        recipient_id:
          type: integer
          title: Recipient Id
          description: ID của đồng chí nhận tin nhắn trực tiếp
      type: object
      required:
      - recipient_id
      title: DirectChatCreate
    DirectiveAckReport:
      properties:
        directive_id:
          type: integer
          title: Directive Id
        recipient_count:
          type: integer
          title: Recipient Count
        acknowledged_count:
          type: integer
          title: Acknowledged Count
        acknowledged:
          items:
            $ref: '#/components/schemas/DirectiveAckUser'
          type: array
          title: Acknowledged
        pending:
          items:
            $ref: '#/components/schemas/DirectiveAckUser'
          type: array
          title: Pending
      type: object
      required:
      - directive_id
      - recipient_count
      - acknowledged_count
      - acknowledged
      - pending
      title: DirectiveAckReport
    DirectiveAckUser:
      properties:
        user_id:
          type: integer
          title: User Id
        full_name:
          type: string
          title: Full Name
        role:
          type: integer
          title: Role
        acknowledged_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acknowledged At
      type: object
      required:
      - user_id
      - full_name
      - role
      title: DirectiveAckUser
...
    DirectiveCreate:
      properties:
        title:
          type: string
          maxLength: 255
          title: Title
        content:
          type: string
          title: Content
        status:
          type: string
          enum:
          - nhap
          - da_ban_hanh
          title: Status
          default: nhap
        classification:
          type: string
          enum:
          - cong_khai
          - noi_bo
          - mat
          title: Classification
          default: noi_bo
      type: object
      required:
      - title
      - content
      title: DirectiveCreate
...
    DirectiveOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        status:
          type: string
          enum:
          - nhap
          - da_ban_hanh
          title: Status
        classification:
          type: string
          enum:
          - cong_khai
          - noi_bo
          - mat
          title: Classification
        author_id:
          type: integer
          title: Author Id
        author_full_name:
          type: string
          title: Author Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
        recipient_count:
          type: integer
          title: Recipient Count
        acknowledged_count:
          type: integer
          title: Acknowledged Count
        acknowledged_by_me:
          type: boolean
          title: Acknowledged By Me
      type: object
      required:
      - id
      - title
      - content
      - status
      - classification
      - author_id
      - author_full_name
      - created_at
      - recipient_count
      - acknowledged_count
      - acknowledged_by_me
      title: DirectiveOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer